Move UrlField to compatibility package
