+++ touchkit-compatibility/src/main/java/org/vaadin/touchkit/v7/gwt/client/vcom/UrlFieldConnector.java
@@ -1,13 +1,13 @@
-package org.vaadin.touchkit.gwt.client.vcom;
+package org.vaadin.touchkit.v7.gwt.client.vcom;
 
-import org.vaadin.touchkit.gwt.client.ui.UrlFieldWidget;
+import org.vaadin.touchkit.v7.gwt.client.ui.UrlFieldWidget;
 
 import com.google.gwt.core.shared.GWT;
 import com.google.gwt.user.client.ui.Widget;
 import com.vaadin.shared.ui.Connect;
 import com.vaadin.v7.client.ui.textfield.TextFieldConnector;
 
-@Connect(org.vaadin.touchkit.ui.UrlField.class)
+@Connect(org.vaadin.touchkit.v7.ui.UrlField.class)
 public class UrlFieldConnector extends TextFieldConnector {
 
     @Override
